+++ packages/yourloops/lib/team/hook.tsx
@@ -288,6 +288,16 @@ function TeamContextImpl(teamAPI: TeamAPI, directShareAPI: DirectShareAPI): Team
     return typeof tm === "undefined";
   };
 
+  const isUserInvitationPending = (user: TeamUser, teamId: string): boolean => {
+    const tm = user.members.find((tm) => tm.team.id === teamId && tm.status === UserInvitationStatus.pending);
+    return !!tm;
+  };
+
+  const isInAtLeastATeam = (user: TeamUser): boolean => {
+    const tm = user.members.find((tm) => tm.status === UserInvitationStatus.accepted);
+    return !!tm;
+  };
+
   const isInTeam = (user: TeamUser, teamId: string): boolean => {
     const tm = user.members.find((tm) => tm.team.id === teamId);
     return typeof tm === "object";
@@ -553,6 +563,8 @@ function TeamContextImpl(teamAPI: TeamAPI, directShareAPI: DirectShareAPI): Team
     isUserTheOnlyAdministrator,
     isInvitationPending,
     isOnlyPendingInvitation,
+    isUserInvitationPending,
+    isInAtLeastATeam,
     isInTeam,
     invitePatient,
     inviteMember,