+++ packages/yourloops/lib/team/models.ts
@@ -151,6 +151,17 @@ export interface TeamContext {
    */
   isOnlyPendingInvitation: (user: TeamUser) => boolean;
   /**
+   * @param user The user to test
+   * @param teamId A team id
+   * @returns {boolean} True if members status is pending in given team
+   */
+   isUserInvitationPending: (user: TeamUser, teamId: string) => boolean;
+  /**
+   * @param user The user to test
+   * @returns {boolean} True if members status is accepted in at least a team
+   */
+   isInAtLeastATeam: (user: TeamUser) => boolean;
+  /**
    * Return true if this user is in a specific team
    * @param user The user to test
    * @param teamId A team id